Extracted the logic for handling the outcome's result for Redis into a separate service.
