+++ Services/Infrastructure/OJS.Services.Infrastructure/Cache/Implementations/CacheService.cs
@@ -8,6 +8,7 @@ using Microsoft.Extensions.Caching.Distributed;
 using OJS.Common.Extensions.Strings;
 using OJS.Services.Infrastructure.Constants;
 using OJS.Services.Infrastructure.ResilienceStrategies;
+using StackExchange.Redis;
 using static OJS.Services.Infrastructure.Constants.ResilienceStrategyConstants.RedisCircuitBreakerOperations;
 
 public class CacheService : ICacheService
@@ -15,21 +16,26 @@ public class CacheService : ICacheService
     private readonly IDistributedCache cache;
     private readonly IResilienceStrategiesService resilienceStrategiesService;
     private readonly IDatesService datesService;
+    private readonly IConnectionMultiplexer redisConnection;
 
     public CacheService(
         IDistributedCache cache,
         IResilienceStrategiesService resilienceStrategiesService,
-        IDatesService datesService)
+        IDatesService datesService,
+        IConnectionMultiplexer redisConnection)
     {
         this.cache = cache;
         this.resilienceStrategiesService = resilienceStrategiesService;
         this.datesService = datesService;
+        this.redisConnection = redisConnection;
     }
 
     public async Task<T> Get<T>(string cacheId, Func<Task<T>> getItemCallback, DateTime absoluteExpiration)
-        => await this.resilienceStrategiesService.ExecuteWithCircuitBreaker(
+        => await this.resilienceStrategiesService.ExecuteRedisWithCircuitBreaker(
             async (_) =>
             {
+                this.CheckRedisConnection();
+
                 await this.VerifyValueInCache(
                     cacheId,
                     getItemCallback,
@@ -47,9 +53,11 @@ public class CacheService : ICacheService
         => this.Get(cacheId, getItemCallback, this.datesService.GetAbsoluteExpirationBySeconds(cacheSeconds));
 
     public async Task Remove(string cacheId)
-        => await this.resilienceStrategiesService.ExecuteWithCircuitBreaker(
+        => await this.resilienceStrategiesService.ExecuteRedisWithCircuitBreaker(
             async (_) =>
             {
+                this.CheckRedisConnection();
+
                 await this.cache.RemoveAsync(cacheId);
                 return Task.CompletedTask;
             },
@@ -82,6 +90,14 @@ public class CacheService : ICacheService
                 cacheId,
                 ParseValue(result),
                 options);
+        }
+    }
+
+    private void CheckRedisConnection()
+    {
+        if (this.redisConnection is not { IsConnecting: false, IsConnected: true })
+        {
+            throw new RedisConnectionException(ConnectionFailureType.UnableToConnect, "Connection to Redis failed. Please check if the Redis server is running and accessible.");
         }
     }
 }
